started defining mvc controllers

+++ src/SuperStore.Application/OutputModels/ProductOutputModel.cs
@@ -7,7 +7,7 @@ public sealed class ProductOutputModel(Product product)
     public string Name { get; } = product.Name;
     public string Description { get; } = product.Description;
     public decimal Price { get; } = product.Price;
-    public decimal Quantity { get; } = product.Quantity;
+    public int Quantity { get; } = product.Quantity;
     public string Category { get; } = product.Category.Name;
     public DateTimeOffset CreatedOn { get; } = product.CreatedOn;
     public DateTimeOffset UpdatedOn { get; } = product.UpdatedOn;
